# feat(weave): allow disabling autopatch