## Description

Gives users the ability to disable autopatching, giving improved `weave.init` performance by almost a second if the user does not need integrations. An example use: a script that doesn't need to log any traces, like the benchmarking script included here.

For discussion related to https://github.com/wandb/weave/issues/3441

Before: Can disable individual integrations

After: Can also easily disable all integrations

Future?: Can selectively enable integrations

Integrations enabled:
```
python benchmark_init.py 
Logged in as Weights & Biases user: jamie-rasmussen.
View Weave data at https://wandb.ai/jamie-rasmussen/2025-01-20_benchmark_init/weave
1.981088749999799
1.254892750000181
1.242289083000287
1.322170541999185
1.3297177079994071
1.2651509999996051
1.3684602919993267
1.268961792000482
1.3221507079997536
1.3851784170001338
  Running init tests... ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 100% 0:00:00 0:00:36


 Init Time Benchmark Results  
┏━━━━━━━━━━━━━━━━━━┳━━━━━━━━━┓
┃ Metric           ┃ Value   ┃
┡━━━━━━━━━━━━━━━━━━╇━━━━━━━━━┩
│ Mean init time   │ 1.3740s │
│ Median init time │ 1.3222s │
│ Std dev          │ 0.2187s │
│ Min time         │ 1.2423s │
│ Max time         │ 1.9811s │
└──────────────────┴─────────┘


  Individual Init Times   
┏━━━━━━━┳━━━━━━━━━━━━━━━━┓
┃ Run # ┃ Time (seconds) ┃
┡━━━━━━━╇━━━━━━━━━━━━━━━━┩
│ 1     │ 1.9811         │
│ 2     │ 1.2549         │
│ 3     │ 1.2423         │
│ 4     │ 1.3222         │
│ 5     │ 1.3297         │
│ 6     │ 1.2652         │
│ 7     │ 1.3685         │
│ 8     │ 1.2690         │
│ 9     │ 1.3222         │
│ 10    │ 1.3852         │
└───────┴────────────────┘
```

Integrations disabled:
```
python benchmark_init.py --disable-autopatch
Logged in as Weights & Biases user: jamie-rasmussen.
View Weave data at https://wandb.ai/jamie-rasmussen/2025-01-20_benchmark_init/weave
0.45323512499999197
0.48261237500082643
0.467393875000198
0.540551707999839
0.4925498749998951
0.44786283299981733
0.4837395419999666
0.4726712920000864
0.38254491699990467
0.5333322499991482
  Running init tests... ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 100% 0:00:00 0:00:25


 Init Time Benchmark Results  
┏━━━━━━━━━━━━━━━━━━┳━━━━━━━━━┓
┃ Metric           ┃ Value   ┃
┡━━━━━━━━━━━━━━━━━━╇━━━━━━━━━┩
│ Mean init time   │ 0.4756s │
│ Median init time │ 0.4776s │
│ Std dev          │ 0.0447s │
│ Min time         │ 0.3825s │
│ Max time         │ 0.5406s │
└──────────────────┴─────────┘


  Individual Init Times   
┏━━━━━━━┳━━━━━━━━━━━━━━━━┓
┃ Run # ┃ Time (seconds) ┃
┡━━━━━━━╇━━━━━━━━━━━━━━━━┩
│ 1     │ 0.4532         │
│ 2     │ 0.4826         │
│ 3     │ 0.4674         │
│ 4     │ 0.5406         │
│ 5     │ 0.4925         │
│ 6     │ 0.4479         │
│ 7     │ 0.4837         │
│ 8     │ 0.4727         │
│ 9     │ 0.3825         │
│ 10    │ 0.5333         │
└───────┴────────────────┘

```