@jamie-rasmussen 2025-01-21
> Even better would be deferring autopatch until the first import so we don't even need this!

That would be better but also a lot more complicated to implement.